refactor: sepreate rpc/rebirth api template to independent files
